Refactored reward kyc actions
